+++ engine/renderer/index_buffer.h
@@ -3,20 +3,20 @@
 
 class IndexBuffer final {
 public:
-	IndexBuffer(uint32_t size);
-	IndexBuffer(const uint32_t* indices, uint32_t count);
+	IndexBuffer(uint64_t size);
+	IndexBuffer(const uint32_t* indices, uint64_t count);
 	~IndexBuffer();
 
 	void bind() const;
 	void unbind() const;
 
-	void set_data(const void* data, uint32_t size) const;
+	void set_data(const void* data, uint64_t size) const;
 
-	uint32_t get_count() const;
+	uint64_t get_count() const;
 
 private:
 	uint32_t ibo;
-	uint32_t count;
+	uint64_t count;
 };
 
 #endif
