+++ engine/renderer/uniform_buffer.h
@@ -3,10 +3,10 @@
 
 class UniformBuffer final {
 public:
-	UniformBuffer(uint32_t size, uint32_t binding);
+	UniformBuffer(uint64_t size, uint32_t binding);
 	~UniformBuffer();
 
-	void set_data(const void* data, uint32_t size, uint32_t offset = 0);
+	void set_data(const void* data, uint64_t size, uint32_t offset = 0);
 
 private:
 	uint32_t ubo;
